requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
attrs==20.2.0             # via pytest
certifi==2020.6.20        # via requests
chardet==3.0.4            # via requests
coreapi==2.3.3            # via drf-yasg
coreschema==0.0.4         # via coreapi, drf-yasg
dj-database-url==0.5.0    # via -r requirements.in
django-choices==1.7.1     # via -r requirements.in
django-filter==2.3.0      # via -r requirements.in
django==2.2.16            # via -r requirements.in, django-choices, django-filter, djangorestframework, drf-yasg
djangorestframework==3.11.1  # via -r requirements.in, drf-yasg
drf-yasg==1.17.1          # via -r requirements.in
factory-boy==3.0.1        # via -r requirements.in
faker==4.1.2              # via factory-boy
idna==2.10                # via requests
inflection==0.5.1         # via drf-yasg
iniconfig==1.0.1          # via pytest
itypes==1.2.0             # via coreapi
jinja2==2.11.2            # via coreschema
markupsafe==1.1.1         # via jinja2
more-itertools==8.5.0     # via pytest
mysqlclient==2.0.1        # via -r requirements.in
packaging==20.4           # via drf-yasg, pytest
pluggy==0.13.1            # via pytest
py==1.9.0                 # via pytest
pyparsing==2.4.7          # via packaging
pytest-django==3.9.0      # via -r requirements.in
pytest==6.0.2             # via -r requirements.in, pytest-django
python-dateutil==2.8.1    # via faker
pytz==2020.1              # via django
requests==2.24.0          # via coreapi
ruamel.yaml.clib==0.2.2   # via ruamel.yaml
ruamel.yaml==0.16.12      # via drf-yasg
six==1.15.0               # via django-choices, drf-yasg, packaging, python-dateutil
sqlparse==0.3.1           # via django
text-unidecode==1.3       # via faker
toml==0.10.1              # via pytest
uritemplate==3.0.1        # via coreapi, drf-yasg
urllib3==1.25.10          # via requests
